BOL-375 refactor: group 관련 model(in data module) class file 분리
