+++ data/src/main/java/com/yapp/bol/data/model/group/GroupListItem.kt
@@ -1,12 +1,10 @@
 package com.yapp.bol.data.model.group
 
-data class GroupDetailResponse(
-    val id: Long,
+data class GroupListItem(
+    val id: Int,
     val name: String,
     val description: String,
     val organization: String,
     val profileImageUrl: String,
-    val accessCode: String,
-    val memberCount: Int,
-    val owner: OwnerDTO
+    val memberCount: Int
 )
