+++ data/src/main/java/com/yapp/bol/data/model/group/GroupSearchApiResponse.kt
@@ -4,12 +4,3 @@ data class GroupSearchApiResponse(
     val content: List<GroupListItem>,
     val hasNext: Boolean
 )
-
-data class GroupListItem(
-    val id: Int,
-    val name: String,
-    val description: String,
-    val organization: String,
-    val profileImageUrl: String,
-    val memberCount: Int
-)
