+++ data/src/main/java/com/yapp/bol/data/model/group/OwnerDTO.kt
@@ -1,0 +1,8 @@
+package com.yapp.bol.data.model.group
+
+data class OwnerDTO(
+    val id: Long,
+    val role: String,
+    val nickname: String,
+    val level: Int
+)
